requirements.txt
Django==4.0.5
colorama==0.4.5
psycopg2-binary==2.9.3
djangorestframework==3.13.1
django-cors-headers==3.13.0
Pillow==9.2.0
django_crontab==0.7.1
cryptography==37.0.4
django-next-prev==1.1.0
python-dateutil==2.8.2
django-user-agents==0.4.0
openpyxl==3.0.10
requests==2.28.2
pandas==2.0.1
django-cleanup==8.0.0
